Bug correction + simplify code

+++ Project02Eclipse/src/com/hemebiotech/analytics/writer/SymptomWriter.java
@@ -2,22 +2,19 @@ package com.hemebiotech.analytics.writer;
 
 import java.io.FileWriter;
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.TreeMap;
 
 public class SymptomWriter implements ISymptomWriter{
 
-    private final String fileName; // Name of final file.
+    private final String fileName;
 
-    /**
-     *
-     * @param fileName a full or partial path to file with symptom strings in it, one per line.
-     */
     public SymptomWriter(String fileName) {
         this.fileName = fileName;
     }
 
     @Override
-    public void write(TreeMap<String, Integer> symptomsMap) {
+    public void write(TreeMap<String, Long> symptomsMap) {
         try {
             FileWriter writer = new FileWriter(this.fileName);
             for (String symptoms : symptomsMap.keySet()) {
